Upload for Sources.
Temporary upload of the sources for V1. Once updated, they will be available in the releases section for this version.

I have included both a 'cleaned' and 'uncleaned' version of the code, the 'uncleaned' version contains old comments, so it may seem messy. The newer, 'cleaned' version has revamped comments.